Criação do FormatUtils e ajuste de identação no mesmo
